Fix multiple pop-ups
Sometimes, multiple pop-up windows were produced for documents in
iframes. In part because the message to enable to add-on was received in
both the top window and the iframe and setup was done in both.

This change fixes a fault in hiding the pop-up (dereferencing an
undefined value if the pop-up is not visible).

This change makes setup to receive messages (onMessage) conditional on
the window being the top level window.