Hide the cursor
To select text with the cursor, the cursor must be on the text. This
makes the text hard to read. This change hides the cursor when it is
selected. The selection is enough to know where the cursor is and hiding
the cursor makes it much easier to read the text.