Fix pop-up position for form elements
In form elements, matched text is not highlighted for some unknown
reason. In this case, the bounds of the matched text were not
determined. This change calculates the bounds of the matched text even
for form elements, but only updates the selection if the element is not
a form element.