Make initialization more robust
Occasionally a freshly loaded iframe doesn't get the addon config.

Failure is intermittent. Most of the time it works as expected.
Presumably there is some sort of timing issue. A race between the tab
adding the onMessage listener and the background script sending the
message? I don't know. The details of when the tab update to complete
occurs is not well defined and I can't be bothered to read the source.

Instead of depending on this, the last thing the content script does in
its foreground is send a 'loaded' message to the background. The
response to this now includes the add-on configuration, which is what
the 'enable' message provides. This response can't come before the
content script is ready to receive it.

So far, so good. But it is an intermittent fault.